move progress util tests
